Add option to vary the mean value passed to random.gauss() call
The mean value is generated using random.triangular(lowerbound,
upperbound) function.
It has been observed that this reduces the height of the bell
curve, substantially (in tests where random.gauss was called
with static and varying means thousands of times). There are
some observe-able differences when varying mean value for the
shakes vs. keeping it static. At any rate, this goes in, for
now, at the very least.